packages/api-pathfinding/requirements.txt
flask==2.0.1
flask-cors==3.0.10
gunicorn==20.1.0
google-cloud-storage==2.1.0
